requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
openai==1.58.1
tiktoken==0.8.0
psutil==6.1.1
Flask-Cors==5.0.0
watchdog==3.0.0